requirements.txt
requests
pycryptodome
bcrypt
PyQt5